+++ tests/test_imports.py
@@ -11,9 +11,11 @@ class TestImports:
         import asp_plot.csm_camera
         import asp_plot.processing_parameters
         import asp_plot.scenes
+        import asp_plot.stereo_geometry
         import asp_plot.stereopair_metadata_parser
         import asp_plot.utils
 
+        assert asp_plot.stereo_geometry is not None
         assert asp_plot.utils is not None
         assert asp_plot.stereopair_metadata_parser is not None
         assert asp_plot.scenes is not None
@@ -28,17 +30,18 @@ class TestImports:
         from asp_plot.altimetry import Altimetry
         from asp_plot.bundle_adjust import PlotBundleAdjustFiles, ReadBundleAdjustFiles
         from asp_plot.processing_parameters import ProcessingParameters
-        from asp_plot.scenes import SceneGeometryPlotter, ScenePlotter
+        from asp_plot.scenes import ScenePlotter
         from asp_plot.stereo import StereoPlotter
+        from asp_plot.stereo_geometry import StereoGeometryPlotter
         from asp_plot.stereopair_metadata_parser import StereopairMetadataParser
         from asp_plot.utils import ColorBar, Plotter, Raster
 
+        assert StereoGeometryPlotter is not None
         assert ColorBar is not None
         assert Raster is not None
         assert Plotter is not None
         assert StereopairMetadataParser is not None
         assert ScenePlotter is not None
-        assert SceneGeometryPlotter is not None
         assert ProcessingParameters is not None
         assert ReadBundleAdjustFiles is not None
         assert PlotBundleAdjustFiles is not None
