+++ tests/test_stereo_geometry.py
@@ -1,0 +1,34 @@
+import matplotlib
+import pytest
+
+from asp_plot.stereo_geometry import StereoGeometryPlotter
+
+matplotlib.use("Agg")
+
+
+class TestStereoGeometryPlotter:
+    @pytest.fixture
+    def stereo_geometry_plotter(self):
+        stereo_geometry_plotter = StereoGeometryPlotter(
+            directory="tests/test_data",
+        )
+        return stereo_geometry_plotter
+
+    @pytest.fixture
+    def stereo_geometry_plotter_tiled(self):
+        stereo_geometry_plotter = StereoGeometryPlotter(
+            directory="tests/test_data/tiled_xmls",
+        )
+        return stereo_geometry_plotter
+
+    def test_dg_geom_plot(self, stereo_geometry_plotter):
+        try:
+            stereo_geometry_plotter.dg_geom_plot()
+        except Exception as e:
+            pytest.fail(f"figure method raised an exception: {str(e)}")
+
+    def test_dg_geom_plot_tiled(self, stereo_geometry_plotter_tiled):
+        try:
+            stereo_geometry_plotter_tiled.dg_geom_plot()
+        except Exception as e:
+            pytest.fail(f"figure method raised an exception: {str(e)}")
